Move static class Startup from ns:AmaiSosu.GUI to ns:AmaiSosi. Tweak Context.Infer() to allow three possibilities: Compile, Help, and default to Install. Add Compile UserControl. Add Install UserControl. Name the Grid and StackPanel elements in MainWindow.
