+++ src/AmaiSosu.GUI/MainWindow.xaml
@@ -3,7 +3,6 @@
                       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                      xmlns:amaiSosu="clr-namespace:AmaiSosu"
                       xmlns:gui="clr-namespace:AmaiSosu.GUI"
                       xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
                       mc:Ignorable="d"
@@ -17,7 +16,7 @@
                       ShowTitleBar="False"
                       WindowStyle="None"
                       Icon="/icon.ico">
-    <!-- Window Styling -->
+  <!-- Window Styling -->
     <Window.Background>
         <ImageBrush ImageSource="Resources/background.png"
                     Stretch="UniformToFill"
@@ -37,7 +36,8 @@
     <!-- Main -->
     <Grid>
         <!-- Header -->
-        <StackPanel VerticalAlignment="Top"
+        <StackPanel x:Name="Header"
+                    VerticalAlignment="Top"
                     HorizontalAlignment="Center"
                     Margin="0,155">
             <Label Foreground="White"
@@ -53,7 +53,9 @@
         </StackPanel>
 
         <!-- Body -->
-        <Grid VerticalAlignment="Bottom"
+        <Grid x:Name="Install"
+              Visibility="{Binding }"
+              VerticalAlignment="Bottom"
               Margin="10,140">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition />
@@ -85,7 +87,8 @@
         </Grid>
 
         <!-- Footer -->
-        <StackPanel Orientation="Horizontal"
+        <StackPanel x:Name="Footer"
+                    Orientation="Horizontal"
                     VerticalAlignment="Bottom"
                     HorizontalAlignment="Center"
                     Margin="0,5">
