+++ src/AmaiSosu/AmaiSosu.csproj
@@ -62,6 +62,7 @@
     <Compile Include="Installation\OpenSauceException.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="Main.Install.cs" />
+    <Compile Include="Main.Startup.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\FileNames.Designer.cs">
       <AutoGen>True</AutoGen>
